Koordinaten Int zu Float geändert.

+++ Client/src/main/java/model/user/IUserController.java
@@ -110,5 +110,5 @@ public interface IUserController {
      * @param posY Y-Koordinate.
      * @see Location
      */
-    void setPosition(int posX, int posY);
+    void setPosition(float posX, float posY);
 }
